Fix almost all clang warnings

+++ src/bricklink/picture.cpp
@@ -44,7 +44,7 @@ int BrickLink::Picture::cost() const
     if (m_image.isNull())
         return 640*480*4 / 1024;      // ~ 640*480 32bpp
     else
-        return m_image.sizeInBytes() / 1024;
+        return int(m_image.sizeInBytes() / 1024);
 }
 
 QFile *BrickLink::Picture::readFile() const
